✨ switch update brew dump

+++ fetchChange.sh
@@ -19,6 +19,18 @@ rm -rf ./.config/karabiner
 mkdir -p ./.config/karabiner
 cp -a ~/.config/karabiner/ ./.config/karabiner/
 
-rm -f .Brewfile
-brew bundle dump -f --global
-cp -a ~/.Brewfile ./.Brewfile
+
+has_brew_option=false
+for arg in "$@"; do
+    if [ "$arg" = "--brew" ]; then
+        has_brew_option=true
+        break
+    fi
+done
+
+if $has_brew_option; then
+    echo "dump brew packages"
+    rm -f .Brewfile
+    brew bundle dump -f --global
+    cp -a ~/.Brewfile ./.Brewfile
+fi
